refactor: extract AbstractReview @MappedSuperclass to fully resolve duplication gate
ProductReview and SellerReview still had byte-for-byte identical
rating/comment/timestamp/visibility fields and @PrePersist/@PreUpdate
callbacks after the ReviewRatingUtils fix, leaving SonarCloud's
new_duplicated_lines_density at 5.3% (still above the 3% max).
Extracted a @MappedSuperclass base with @SuperBuilder (not plain
@Builder, which doesn't include inherited fields) so both entities'
existing .builder()...build() call sites keep working unchanged.
